requirements.txt
aiohttp
aiohttp-retry
aiosignal
asgiref
attrs
certifi
cffi
charset-normalizer
cryptography
Django
django-dotenv
djangorestframework
frozenlist
idna
jwt
multidict
pycparser
PyJWT
python-dotenv
pytz
razorpay
requests
sqlparse
stripe
twilio
typing_extensions
urllib3
yarl
